Update patch set 1
Patch Set 1:

The source of this issue is in https://review.openstack.org/327503, which uses code from a newer version of dogpile.cache (0.6.0 or later). That change should have depended on https://review.openstack.org/327297.

In any case, we need to have a newer version of python-dogpile-cache in the requirements repo.

Patch-set: 1